Module Types.FEATURE

Optional features of COBOL dialects with support levels similar to those in GnuCOBOL.

type 'a support_level =

| Ok of 'a

| Warning of 'a

| Archaic of 'a

| Obsolete of 'a

| Skip

| Ignore

| Error

| Unconformable

Values for support levels. Those that do not raise errors may come with option-specific values, hence the type parameter.

class virtual 'a feature_kind : name:string -> object ... end

Kind (type) of support level options

val unit_kind : unit feature_kind

Generic kind of support, that carry no value; i.e, those are simple flags.

type 'a verification_result = ('a * (Cobol_common.srcloc * 'a support_level configurable * [ `Archaic of configuration | `Obsolete of configuration | `Used ]) option, (Cobol_common.srcloc * 'a support_level configurable * [ `Ignored | `Unconformable of configuration | `Used ]) option) Stdlib.result

class 'a support : 'a0 support_level configurable -> config:configuration -> 'a1 support_level -> object ... end

Internal representation of a binding from a feature and a support level.

val verify : 'a support -> loc:Cobol_common.srcloc -> 'a verification_result

class 'a feature : feature_kind:'a0 feature_kind -> name:string -> ?short:doc option -> doc -> object ... end

Type of optional COBOL dialect features.

type 'a t = 'a feature

val def : feature_kind:'a feature_kind -> name:string -> ?short:doc -> doc -> 'a feature

Definitions

val unit : name:string -> ?short:doc -> doc -> unit feature